Component/containers refactoring for tasks page
